+++ src/addons/display_commander/settings/advanced_tab_settings.cpp
@@ -52,7 +52,6 @@ AdvancedTabSettings::AdvancedTabSettings()
       suppress_wgi_for_non_unity_games("SuppressWgiForNonUnityGames", false, "DisplayCommander"),
       debug_break_on_severity("DebugBreakOnSeverity", false, "DisplayCommander"),
       suppress_window_changes("SuppressWindowChanges", false, "DisplayCommander.Safemode"),
-      enable_dxgi_refresh_rate_vrr_detection("EnableDxgiRefreshRateVrrDetection_default_on", true, "DisplayCommander"),
       win_up_grace_seconds("WinUpGraceSeconds", 1, 0, 61, "DisplayCommander"),
       disable_dpi_scaling("DisableDpiScaling", true, "DisplayCommander") {}
 
@@ -91,7 +90,7 @@ std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetAllSettings() {
             &enable_performance_overlay_shortcut,
             &suppress_wgi_globally, &suppress_wgi_enabled, &suppress_wgi_for_unity, &suppress_wgi_for_non_unity_games,
             &debug_break_on_severity, &suppress_window_changes,
-            &enable_dxgi_refresh_rate_vrr_detection, &win_up_grace_seconds, &disable_dpi_scaling};
+            &win_up_grace_seconds, &disable_dpi_scaling};
 }
 
 std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetSettingsToSave() {
@@ -107,7 +106,6 @@ std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetSettingsToSave() {
             &suppress_wgi_for_non_unity_games,
             &debug_break_on_severity,
             &suppress_window_changes,
-            &enable_dxgi_refresh_rate_vrr_detection,
             &disable_dpi_scaling};
 }