+++ src/addons/display_commander/ui/new_ui/debug/dxgi_refresh_rate_tab.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "../../../ui/imgui_wrapper_base.hpp"
-
-namespace ui::new_ui::debug {
-
-void DrawDxgiRefreshRateTab(display_commander::ui::IImGuiWrapper& imgui);
-
-}  // namespace ui::new_ui::debug